requirements.txt
autoenv
Flask
Jinja2
psycopg2-binary
python-dotenv
SQLAlchemy
WTForms
